hgmo: serve stream clone bundles to private IP address on cloud instances (Bug 1518394) r=smacleod
When Taskcluster hosts pull/clone from private hgweb mirrors in CI,
they do so over a VPC peering connection and do not cross the public
internet. As a result, the IP address exposed to the hg process on
the mirror host sees a private IP address as the source location. This
confuses the clonebundles manifest sorting function, as that function
reads a previously fetched object containing AWS IP blocks for Amazon's
public IP address space. Since the private IPs do not appear in the
object, pulling from the mirror results in the Taskcluster hosts
receiving zstd-max compressed bundles, instead of the faster streamed
clone bundles.

This commit teaches the hgmo extension to serve stream clone bundles
when the client IP address is within the private IP address space and
when the `instance_data.json` object is present on the server. The
`instance_data.json` object contains some useful information gathered
about the instance during first run by cloud-init, including the region
of the host mirror. We grab the region from this object and pass it
to `filter_manifest_for_aws_region` to achieve the same behaviour as
if we received a result from the public AWS IP space.

Differential Revision: https://phabricator.services.mozilla.com/D16216

--HG--
extra : moz-landing-system : lando