hgmo: get `Remote-Addr` header from `rawenv` (Bug 1518394) r=smacleod
The Zeus load balancer sets an `X-Cluster-Client-Ip` header
to any request for hg.mo. The alternative header we use to
identify the source IP address, `Remote-Addr`, is set as an
environment variable to the WSGI process that spawns to handle
the HTTP request. The mismatch in locations is due to the
way Mercurial parses incoming HTTP requests - only choosing
to recognize headers beginning with `HTTP_` for inclusion
in the `mercurial.hgweb.wgsiheaders.Headers` object. The
other values set as environment variables are still available
in the `rawenv` container, so this commit switches the value
of `sourceip` to come from that object when required. The
corresponding test is updated to reflect the correct behaviour.

Since http requests to docker images running on a local machine
come from private IP addresses, and the docker images are built
with an `instance_data.json` object into them starting with
commit 09fa6119e23b, making this change causes all requests
in the test suite to default to streamed clone bundles. This
commit also renames the instance data object to a non-standard
name during docker image builds and adds a helper function to
toggle test features for cloud based instances. For now this
simply means renaming the file back to it's standard location.

Differential Revision: https://phabricator.services.mozilla.com/D16569

--HG--
rename : ansible/roles/docker-hg-web/files/instance_data.json => ansible/roles/docker-hg-web/files/test_instance_data.json
extra : moz-landing-system : lando