ansible/hg-web: add `replicationrules` to vcsreplicator config if target host is a mirror (Bug 1511058) r=glob
When we deploy the `hg-web` role to mirrors in AWS, we
want to apply replication filters to only transfer
data about the repositories relevant to CI. This commit
teaches the hg-web role to add a pre-defined
`replicationrules` config section to the vcsreplicator.ini
file on the new host, when a `replication_filters`
Ansible variable is set to `True`. We default to `False`
for the hg-web role as we consider our on-premesis
hosts to be the default.

Differential Revision: https://phabricator.services.mozilla.com/D16633

--HG--
extra : moz-landing-system : lando