ansible: add hgweb mirrors to hgmo deployment playbook (Bug 1511058) r=glob
This commit adds the hgweb mirror hosts to the hgmo deployment
playbook. Mirror hosts are applied the hg-web role with the
`replication_filters` variable set to `True`. Similarly
to previous commits, we add a flag to the `./deploy` script to
be able to skip deploying to the mirror hosts.

Differential Revision: https://phabricator.services.mozilla.com/D16635

--HG--
extra : moz-landing-system : lando