ansible: add a playbook to deploy hgweb mirrors in AWS (Bug 1511058) r=glob
This commit adds a `bootstrap-hgweb.yml` playbook to run the
vcsreplicator bootstrap procedure on hosts in a special Ansible
inventory group. The play runs the hgssh component of the bootstrap,
which creates entries in the replication log for each repo on hg.mo,
and produces a JSON artifact with information about the repositories.
That artifact is transferred by Ansible to the hosts requiring
bootstrap, where it is ingested by the hgweb component.

This playbook includes a workaround for a performance issue with
the try repo. Since vcsreplicator works by pulling individual
changesets (as opposed to a clone), the try repos take far too
long to replicate in the bootstrap procedure. To get around this
issue, we comment out the try repository replication rule before
running the hgweb bootstrap and stream clone the repo over HTTPS.
We then re-enable the try repos for replication once the clone is
complete and restart the replication daemons.

This playbook is able to be used both as a standalone play, or
as a separate step in the hgmo deploy playbook. A future commit
will teach the deploy script to run this playbook in the presence
of a flag.

Differential Revision: https://phabricator.services.mozilla.com/D16634

--HG--
extra : moz-landing-system : lando