robustcheckout: mark as compatible with Mercurial 4.9 and update tests (Bug 1513273) r=smacleod
This commit marks robustcheckout as compatible with Mercurial
4.9 and updates a failing test with conditional output on the
new version. The new output in the tests comes from the addition
of a new `sparserevlog` repo requirement[0]. Essentially this
new line in the test output indicates that the repo in question
was created with a new version of Mercurial and the same or
newer version of Mercurial must be used to act upon it. This
only applies to repositories accessed on the same filesystem,
as any legacy version of Mercurial can clone/pull the repo
over the wire and the resulting repo on disk will not have this
requirement.

Bug 1513276 tracks the work of upgrading repos on hgmo to the
new storage format.

[0] https://www.mercurial-scm.org/wiki/MissingRequirement

Differential Revision: https://phabricator.services.mozilla.com/D17243

--HG--
extra : moz-landing-system : lando