terraform: add Taskcluster VPC to subnet route tables (Bug 1516029) r=bstack
This commit adds a route to send traffic from the Taskcluster
CIDR block to the peering connection interface between our
VPC and the Taskcluster VPC. The block is passed as a parameter
to the `ci-only` Terraform module and then sent to each private
subnet route table as needed. Since managing the resource from
within Terraform would require us to either store non-public
information about the Taskcluster VPC (such as the ID and account
number) in this public repo, or maintain a separate document/service
with the correct details that can be fed into Terraform, we
instead gather the information about the connection using a
data source and search for the resource tagged with the
"Name" key "Taskcluster-hg".

Differential Revision: https://phabricator.services.mozilla.com/D16852

--HG--
extra : moz-landing-system : lando