terraform: remove bastion hosts and allow access to mirrors from MozVPN (Bug 1509039) r=bstack
Now that SSH access to the VPC is limited to MozVPN, there is
no need for the bastion host we were previously using. This
commit removes the bastion module and all associated resources,
and updates the mirrors security group to allow SSH from MozVPN
through the VPN gateway.

Differential Revision: https://phabricator.services.mozilla.com/D17652

--HG--
extra : moz-landing-system : lando