ansible: make `monitoring-agent` role a dependency of `hg-web` role (Bug 1525954) r=glob
This commit adds the previously created `monitoring-agent` role as
a dependency of the `hg-web` role. This will cause the Telegraf
monitoring agent to be installed on all hg-web hosts (AWS mirrors
and on-premises hosts) and report metrics to the InfluxDB instance
running in AWS.

I have deployed the agent to the mirror hosts as a test already
and things work as intended. Once I deploy to the main datacentre
hosts I anticipate that I will need to request firewall access
to get traffic flowing through the VPN gateway. In the event this
is true I will simply deactivate the Telegraf service and restart
it once the access is granted.

Differential Revision: https://phabricator.services.mozilla.com/D20359

--HG--
extra : moz-landing-system : lando