ansible: add `monitoring-agent` Ansible role (Bug 1525954) r=glob
This commit adds a `monitoring-agent` Ansible role which will
install the Telegraf reporting agent from the InfluxDB stack
on desired hosts. As a basic first step, in this commit we install
the RPM package and install a basic config file which enables
all plugins required to populate the pre-canned Chronograf dashboards
to monitor Apache and system resources.

Installing Telegraf in this manner will create a `telegraf` user
on the host and load a Telegraf unit into systemd.

Differential Revision: https://phabricator.services.mozilla.com/D20358

--HG--
extra : moz-landing-system : lando