hgext/clang-format: wrap `cmdutil.amend` to run clang-format during `hg commit --amend` (Bug 1525725) r=sylvestre
This commit adds support for `hg commit --amend` to the clang-format
extension. `cmdutil.amend` is wrapped and some of the basic logic
(such as building the matcher object and creating the list of files
to be amended) is pulled into the wrapper function to determine if
clang-format should be applied.

`cmdutil.amend` should be safe to wrap here since it is only called
from within the `amend` code branch of the `commit` command.

While we add the `--amend` test to `test-clang-format.t`, we also
add some spacing to other lines in the test to improve readability.

Differential Revision: https://phabricator.services.mozilla.com/D20514

--HG--
extra : moz-landing-system : lando