ansible: use `yum` module to install packages (Bug 1525954) r=smacleod
This commit switches all uses of the `command` module which call
out to `yum` to instead use the `yum` Ansible module directly.

A few locations still call out to `yum` to use the versionlock
yum plugin. I found an open GitHub issue stating there is no way
to accomplish this with the `yum` module directly, so those are
left as-is for the time being.

Differential Revision: https://phabricator.services.mozilla.com/D20682

--HG--
extra : moz-landing-system : lando