ansible/monitoring-host: install Kapacitor to monitoring host (Bug 1525954) r=glob
As a first step towards monitoring hg.mo with the InfluxDB stack,
we installed the database, user interface and monitoring agent
(InfluxDB, Chronograf and Telegraf, respectively). Now that we
have some of the basic dashboards and monitoring data being
collected and queryable, we need to install Kapacitor (the Influx
data processing and alerting tool). Once we have Kapacitor installed
we can port the existing Nagios alerts to TICKscripts, and add
functionality for processing the data via user defined functions
(UDF's).

Differential Revision: https://phabricator.services.mozilla.com/D23170

--HG--
extra : moz-landing-system : lando