ansible: apply `monitoring-agent` role to hgweb mirrors (Bug 1525954) r=glob
We will need this role applied to get data from the
hgweb mirrors. Previously we applied the role as
a dependency of the hg-web role, but that broke
the test suite as the monitoring role assumed
the presence of the `/var/log/httpd` directory on
disk. It was rolled back in f923e988151c to fix
the test suite.

Differential Revision: https://phabricator.services.mozilla.com/D23172

--HG--
extra : moz-landing-system : lando