ansible/monitoring-agent: switch from `systemd_unit` to `pattern` use of `procstat` plugin (Bug 1525954) r=glob
When I attempted to port the vcsreplicator daemon count
alert to Kapacitor/TICKscript, I realized that instructing
the `procstat` plugin to monitor a specific systemd unit
would not tell InfluxDB when the unit failed, only when
a unit of that name was not loaded into the system. By
switching to using the `pattern` syntax (with the value
being a pattern sent to `pgrep`) we can use the plugin
to correctly determine when we do not have the required
amount of vcsreplicator daemon processes running on the
system.

Differential Revision: https://phabricator.services.mozilla.com/D23174

--HG--
extra : moz-landing-system : lando