ansible/monitoring-host: add TICKscript to alert for missing vcsreplicator processes (Bug 1525954) r=glob
This commit takes the first step in porting the existing Nagios
checks to the InfluxDB platform (specifically Kapacitor). We create
an alert which queries the `procstat_lookup` measurement for points
with a `pattern` tag matching "vcsreplicator-consumer". When
we do not find 8 running systemd units with this prefix (ie the
`vcsreplicator@{0..7}.service` units on the hgweb hosts) we trigger
the alert.

Differential Revision: https://phabricator.services.mozilla.com/D23177

--HG--
extra : moz-landing-system : lando