vcsreplicator: add `Telegraf` flag to Nagios checks (Bug 1525954) r=glob
This commit adds a flag to format output from the
custom Nagios checks for consumption by Telegraf.
The `exec` Telegraf plugin will query the offset
checking script with this parameter (as shown in
the test) and send the returned value back to Influx
for use in an alerting TICKscript.

Differential Revision: https://phabricator.services.mozilla.com/D23175

--HG--
extra : moz-landing-system : lando