robustcheckout: add a `serverUrl` key to CI metrics output (Bug 1535011) r=smacleod
With private hg mirrors being rolled out in the near future,
we would like to be able to differentiate our CI metrics
between operations against the public hgmo service, and
the private mirrors. In this commit we parse the url passed
to `robustcheckout` using `urlparse.urlparse`, and add
the network location to the output suites for all data points.

Differential Revision: https://phabricator.services.mozilla.com/D23530

--HG--
extra : moz-landing-system : lando