configwizard: move `curses` availability check to standalone function (Bug 1531059) r=glob
It turns out we already check for the presence of `curses`
on the system in a separate function in `configwizard`,
however the check is coupled with a prompt to enable
the curses interface. Here we move the check itself into
a separate function so we can re-use the functionality
in the next commit.

Differential Revision: https://phabricator.services.mozilla.com/D23707

--HG--
extra : moz-landing-system : lando