configwizard: set `ui.interface=text` on Windows (Bug 1531059) r=glob
Tweakdefaults causes the default interface to change to `curses`
on Windows. This can cause issues as curses is not supported out
of the box. This commit adds a check to ensure `curses` is available
on the user's system and reverts the default UI to `text` when
it is not.

Differential Revision: https://phabricator.services.mozilla.com/D21705

--HG--
extra : moz-landing-system : lando