Bug 1522496 step 1.5 - simplify testing of known exceptions  r=sheehan
Previous versions included tracebacks when a test induced an exception.
This added no value to the test and made it more difficult to move
dependent modules or upgrade to new versions of Python.  The code was
modified to move from tracebacks with unhandled exceptions to just
noting the exception and associated metadata on stdout before aborting.

Removed the ability to execute generate_hg_s3_bundles.py directly in
favor of the entry_point.

Differential Revision: https://phabricator.services.mozilla.com/D24256

--HG--
extra : moz-landing-system : lando