testing: make `run-tests` use the flaky test blacklist by default (Bug 1539585) r=lars
Previously we added a `flaky-tests.txt` list of tests in
version-control-tools which intermittently or consistently
fail. This commit makes `run-tests` respect that list by
default, with the option to run the flaky tests when
`--run-flaky` is passed to the test runner. We will still
need to run the tests for major feature changes, to ensure
that it fails in the same way as before (ie not introducing
any new failure output).

Differential Revision: https://phabricator.services.mozilla.com/D25388

--HG--
extra : moz-landing-system : lando