hgserver: add new bundle2 capability to `test-obsolescence-server.t` (Bug 1539585) r=lars
This new output was introduced in Mercurial 4.7. The server is
now capable of transmitting v2 stream clone bundles, as shown
in this new output. See `hg help internals.wireprotocol.bundle2`
for more information on the format of the new content.

The output is now alphabetically sorted by default, negating the
need for the previous line of unsorted (but deterministic) output.

Differential Revision: https://phabricator.services.mozilla.com/D25385

--HG--
extra : moz-landing-system : lando