hgmo: make compatible with Mercurial 4.9 (Bug 1513273) r=lars
Test extension works as intended under 4.9, with a few
minor additions to tests. A new cache file `wcache` has
been added as a virtual file system cache for files
related to the working copy. The UNIX-like exec flag
check caches are no longer shown when displaying the
caches available on the local repository as the code
path that created them has changed. Considering they
are not actually transferred later in the test, this
is not a problem.

Differential Revision: https://phabricator.services.mozilla.com/D25952

--HG--
extra : moz-landing-system : lando