configwizard: remove warning about experimental status of `evolve` extension (Bug 1541887) r=glob
`evolve` is stable enough that we recommend it to developers
working on Firefox repos. The warning about it's "experimental"
status is confusing and is likely driving people away from
adopting it for their workflow. This commit removes the warning
and indicates that installation of `evolve` is desirable. We
also point developers at the relevant help commands to learn
more for themselves.

Differential Revision: https://phabricator.services.mozilla.com/D26203

--HG--
extra : moz-landing-system : lando