vcsreplicator: update test, correcting order of expected output (Bug 1543448) r=sheehan
Machines of differing speeds produce the output of the test,
test-consumer-daemon-restart.t, in a differing orders. This
change synchronizes the output by adding a consumer
with "--wait-for-no-lag". This causes a pause that allows the
tests that run subsequently to always have the same order
of expected output.

The former order of output has been corrected for this
synchronization and a redundant line about shutdown
has been removed (see Bug 1543450)

Differential Revision: https://phabricator.services.mozilla.com/D27005

--HG--
extra : moz-landing-system : lando