vcsreplicator: simplify shutdown signal handlers (Bug 1543450) r=sheehan
The existing shutdown signal handler fails to take into account the EINTR
exception raised by blocking system calls.  This is raised when the signal
is caught during a system call like reading from a socket by Kafka.  This
happens when the signal handler simply returns without raising an exception
of its own.  The exsting handler used a counter as a closure to control the
message consuming loop.

As an alternate implementation, this new handler raises a custom exception
whenever a signal is received.  This doesn't allow the blocking system call
the opportunity to raise the EINTR exception.  The custom exception breaks
the consuming loop without the need of a counter.

An exception handler was added to the try:finally: construct to take
the responsibility of logging the shutdown process.

For this change to pass tests, Bug 1543448 must also be resolved.

Differential Revision: https://phabricator.services.mozilla.com/D27004

--HG--
extra : moz-landing-system : lando