ansible: turn off services irrelevant to hgweb mirrors during deploy (Bug 1511058) r=glob
We don't need to serve traffic to Mozilla offices
from the private hgmo service, so we can turn off
the IP address scraper service.

In the future we should refactor the hgweb role
to do this automatically (ie not install an unnecessary
service in the first place).

Differential Revision: https://phabricator.services.mozilla.com/D26990

--HG--
extra : moz-landing-system : lando