ansible: fix bootstrap and teach it to apply `hg-web` role (Bug 1511058) r=glob
There were a few subtle bugs with the hgweb bootstrap
playbook, including the fact that it assumed the hg-web
role was already applied. This commit adds code to
optionally apply hgweb/mirror playbooks on the existance
of an appropriate variable. The directory of the
`hgssh.json` file is changed to the `hg` users home.
We set SELinux to permissive mode on the mirror hosts
as `mod_wsgi` cannot be accessed with it turned on.
Finally, we add the bootstrap hostgroup to the `hgmo`
group, to allow access to it's default group variables,
which include the set of Kafka/Zookeeper brokers.

Differential Revision: https://phabricator.services.mozilla.com/D26992

--HG--
extra : moz-landing-system : lando