vcsreplicator: add a test demonstrating pull behaviour on hgssh (Bug 1513273) r=lars
In attempting to fix `test-pushkey-obsolescence.t` to make
vcsreplicator compatible with Mercurial 4.9, I needed to allow
serving the `unfiltered` repoview on hgssh. This had the
potential to change the behaviour of pulls against the ssh
endpoint on hgmo, so to ensure we don't change the intended
behaviour by accident, I've added a test which demonstrates
the current behaviour.

Differential Revision: https://phabricator.services.mozilla.com/D27778

--HG--
extra : moz-landing-system : lando