meta: add `.sops.yaml` and basic documentation for working with `sops` (Bug 1543991) r=lars
As described in the bug, we would like to be able to check in secrets
to version-control-tools for a variety of reasons. This commit
introduces a `.sops.yaml` file and a brief introduction to using
`sops` to encrypt configuration secrets checked in to version control.

For the moment, all secrets in this repo are encrypted with a single
GCP KMS key. The recommended practice is to also encrypt with a
backup PGP key that is kept somewhere offline (like a safe in an
office) for backup purposes. Considering the nature of the items
I want to place in the repo, I don't think we should block on doing
that for now.

Differential Revision: https://phabricator.services.mozilla.com/D28361

--HG--
extra : moz-landing-system : lando