vcsreplicator: add `--telegraf` flag to aggregator lag check (Bug 1545434) r=smacleod
This commit adds a `--telegraf` flag to the aggregator lag check
script running on hgssh. When the script is run with this flag,
the output will be formatted as JSON, consumable by Telegraf's
`exec` plugin. With this flag added, we will be able to send
aggregator lag data to InfluxDB for monitoring/alerting purposes.

Differential Revision: https://phabricator.services.mozilla.com/D28151

--HG--
extra : moz-landing-system : lando