ansible/monitoring-agent: add `hgssh` specific plugins to Telegraf config (Bug 1545434) r=smacleod
This commit adds a `hgssh` specific set of configuration options
to the Telegraf configuration file. At the moment we track the
lag of the "aggregator" replication daemons, and the presence
of processes for those daemons. We also apply the `monitoring-agent`
role in the hgssh deploy playbook and set `hgssh` as the hosttype.

Differential Revision: https://phabricator.services.mozilla.com/D28152

--HG--
extra : moz-landing-system : lando