deployments: teach `./deploy` to decrypt secrets with `sops` (Bug 1543991) r=lars
Now that we can use `sops` to encrypt/decrypt secrets in
version-control-tools, we need to automate the process
of decrypting these secrets when we want to deploy them
to the service. This commit adds a step to the `./deploy`
tool which parses the `.sops.yaml` file for a `files` key,
inside which is a list of files which need to be decrypted
in-place before running through the deployment.

Differential Revision: https://phabricator.services.mozilla.com/D28362

--HG--
extra : moz-landing-system : lando