hgtemplates: wrap instances of `author` template keyword with `mailmap` function (Bug 1464838) r=glob
Mercurial has an implementation of Git's `mailmap` available as a
template function. I added the implementation as a way of easing
my way into learning the Mercurial codebase. This commit performs
a regex replace of `{author` with `{mailmap(author)` everywhere
in `gitweb_mozilla`, and adds the associated patch file.

Differential Revision: https://phabricator.services.mozilla.com/D29549

--HG--
extra : moz-landing-system : lando