ansible/monitoring-agent: add monitoring of Kafka/Zookeeper processes (Bug 1548266) r=glob
We'd like to add monitoring of our Kafka brokers to the
`monitoring-agent` Ansible role. In this commit we create
a new `kafka.conf.j2` template which will include all
Telegraf plugins used to monitor Kafka, and include the
template in our Telegraf config on the condition the
Kafka host is in our `zk_servers` `hgmo` group variable.

Differential Revision: https://phabricator.services.mozilla.com/D29498

--HG--
extra : moz-landing-system : lando