terraform: remove subnets in availability zone C (Bug 1548302) r=bstack
Turns out, us-west-1 doesn't have an availability zone C. This
was mentioned to me during review, and I thought I checked each
region to ensure there was a third AZ for each, but apparently
the way I checked was wrong. Oops!

Considering we aren't using the subnet in that AZ at the moment
any way, this isn't a huge deal. So this commit removes the C
availability zone subnets (public and private).

Differential Revision: https://phabricator.services.mozilla.com/D30293

--HG--
extra : moz-landing-system : lando