terraform: create unique certificate per region (Bug 1548302) r=bstack
The previous plan was to create a single certificate that would
be shared across AWS regions, but we have since learned that
ACM certificates are per-region only. So this commit moves the
certificate resource definition from the `main.tf` file for the
entire Terraform config and into the `aws-vpc` modules for each
specific AWS region. We also remove helper variables from the
module.

Differential Revision: https://phabricator.services.mozilla.com/D30296

--HG--
extra : moz-landing-system : lando