terraform: upgrade mirror instances to `c5d.2xlarge`
The instances were previously set to this type, but I
moved them down as they could handle the traffic.
Unfortunately when project repos like `larch` or `maple`
need to pull, the increased amount of CPU required to
calculate the response multiplied by the number of
responses that must be generated create too much load.