hgext: update deprecated call to ui.progress (Bug 1550175) r=sheehan
For the extension `overlay`, the call to `ui.progress` was just replaced
`ui.makeprogress`.  This is justified by the declaration in `overlay`
requiring HG47 or later.  Since `ui.makeprogress` was introduced in HG47,
this seem justified.

For the extension `mozext`, since I know nothing of what versions of
HG will be used, a local function was added of the form:
`makeprogress(ui, ...)` to the call to `ui.progress`.  In that
method, `ui.makeprogress` is tried.  If that raises an `AttributeError`,
then the original `ui.progress` method is called.

Differential Revision: https://phabricator.services.mozilla.com/D30564

--HG--
extra : moz-landing-system : lando