ansible/hg-ssh: add `autoland` to mozilla-unified (Bug 1551274) r=smacleod
This commit adds the `autoland` repository to the set of repositories
in mozilla-unified. The tip revision of this repository will have
a bookmark named `autoland` at it's tip revision, copying the section
name in the config file. Since all new commits will be on the default
branch, we limit the pulled revisions to that branch.

Differential Revision: https://phabricator.services.mozilla.com/D31115

--HG--
extra : moz-landing-system : lando