terraform: tear down AWS resources in `eu-central-1` and `us-east-2` (Bug 1553885) r=dustin
While these regions have a Gecko workers VPC available,
it appears there are no workers configured to run in them.

In this commit we tear down the hgmo VPC in the same regions.
We can easily stand up the VPCs again if we set up workers in
those regions.

Differential Revision: https://phabricator.services.mozilla.com/D32727

--HG--
extra : moz-landing-system : lando