TICKscripts: add alert for too many try heads (Bug 1556530) r=glob
In the previous commit we added reporting for the number of
heads on try, and in this commit we tell Kapacitor to send a
Slack alert in the event we cross 40000/50000 heads (as a warning
and critical alert, respectively). This script is checked in to
version-control-tools with the other alerting scripts for tracking
purposes but will need to be manually added to Kapacitor through
Chronograf once the previous commit is deployed.

Differential Revision: https://phabricator.services.mozilla.com/D33552

--HG--
extra : moz-landing-system : lando