telegraf: add an `exec` plugin to report the number of heads on try (Bug 1556530) r=glob
This commit adds a Telegraf `exec` input plugin to monitor the
number of heads on the try repo on hgssh. The command is a call
to `hg log`, taking a call to the `revset` template function and
passing it through the `count` template filter to get the number
of revisions in the specified revset (in this case, `head()`).
The data is sent as a single integer to InfluxDB and a follow
up commit will introduce a TICKscript to send alerts when the
number of heads grows too high.

Despite using the `head()` revset from within the template for
the command, we specify `-r 0` as the revset for the log command.
Mercurial's templater applies the given template to each revision
in the specified revset, and since we only want to run the revset
a single time and then count the number off revisions, we specify
a single revision in the `-r` argument.

Differential Revision: https://phabricator.services.mozilla.com/D33551

--HG--
extra : moz-landing-system : lando