hgtemplates: run `test-template-sync.t` against Mercurial 5.0.1 (Bug 1557070) r=glob
This version should always match the version running on hg.mo,
which means it needs to be bumped during the upgrade. Running
this test should tell us if we have any outstanding template
upgrades to perform, and since it passes without issue, we
don't need to make any template changes.

Depends on D33849

Differential Revision: https://phabricator.services.mozilla.com/D33850

--HG--
extra : moz-landing-system : lando