py3: global replace `iteritems()` with `items()` (Bug 1541166) r=lars
`iteritems()` doesn't exist in Py3, with `items()` instead returning
a generator by default. This commit switches all instances of the
old `iteritems` method with `items`. In all cases except one, we
were evaluating the full list of items regardless of the generator
based function call, and the instance where we were only returned
a small number of items, so I don't expect any performance hit here
while we are still on Python 2.

The same operation has been performed for `iterkeys()` and `keys()`,
but only a single instance was found and is included in this commit.

There are still some instances of `iteritems` in the repo, in the
`third_party` directory of vendored libraries. These will be handled
in bug 1557783.

Differential Revision: https://phabricator.services.mozilla.com/D34193

--HG--
extra : moz-landing-system : lando