format-source: enable compatibility with Mercurial 5.0 (Bug 1545236) r=andi
format-source was never made compatible with Mercurial 5.0, since
the extension was disabled on Firefox developer machines. Now
that we plan to use the extension for a JavaScript reformat using
`prettier`, the extension will need to be compatible with newer
versions of Mercurial.

In this commit we apply compatibility patches from upstream
`format-source`[0] to enable 5.0 compatibility, with some
modifications. Rather than perform a stack inspection to check
for the correct function signature to use, we look at the version
of Mercurial being run and call `cmdutil.add` with the correct
signature for each version. We also define a single function
with a conditional inside instead of a global conditional with
two separate function definitions.

--
[0] https://bitbucket.org/octobus/format-source/commits/e526ce24a5e874c2174c5fc6abda7c7163a18e26

Differential Revision: https://phabricator.services.mozilla.com/D34884

--HG--
extra : moz-landing-system : lando