ansible: remove `git` dependency and retrieve `cinnabarclone.py` using `get_url` (Bug 1558868) r=glob
Installing `git` on the hg hosts in MDC1 fails due to a custom YUM repo
added by the RelOps/IT teams. Instead of blocking the deployment of
cinnabarclone on hgmo as a result of this bug, let's pull down a specific
version of `cinnabarclone` using `get_url`, and deploy that.

We make the download of the extension conditional on a different checksum
for the file in question. This will avoid re-downloads when the file already
exists on-disk with the specified checksum.

Differential Revision: https://phabricator.services.mozilla.com/D35288

--HG--
extra : moz-landing-system : lando