push-to-try: check for valid try_task_config.json file. r=sheehan
When using a try_task_config.json file on a try push, try to parse it first
to prevent failures.

I added this based on personal experience using this feature. I usually forget
to check first and then the push fails.

No attempt it made to ensure the file is valid beyond ensuring it is valid
JSON that can be loaded by Python's "json" stdlib module.

Differential Revision: https://phabricator.services.mozilla.com/D35193

--HG--
extra : moz-landing-system : lando