hgmo: remove wire protocol v2 hacks (Bug 1550748) r=smacleod
After landing much of the wire protocol v2 work upstream, we
started finding lots of bugs and patching over them in v-c-t,
with the intent of finding more issues and fix them upstream
after some experimentation. Since there are no plans to continue
working on this for the time being, I'm removing it, as well as
some of the related tests.

Differential Revision: https://phabricator.services.mozilla.com/D36701

--HG--
extra : moz-landing-system : lando