hgtemplates: allow passing a path `hg` executable to `mozify-templates.py` (Bug 1541166) r=smacleod
In a later commit, I will upgrade our copy of the Mercurial test
harness for Python 3 compatibility. In doing so, the template
sync test fails as `mozify-templates` no longer uses the version
of `hg` in the test - instead finding the system `hg` and using that.
Since my version is 4.5.3 on the server and I am typically running
tests against 5.0.1 these days, the tests fail as 4.5.3 does not have
the `sparse-revlog` repository requirement.

This commit adds the ability to specify a specific version of hg
for use in the script, by passing a path in an environment variable.
It's a quick hack to make things work - we should probably add some
option parsing to `mozify-templates.py` at some point, at which time
we can convert the environment variable to an option.

Differential Revision: https://phabricator.services.mozilla.com/D36703

--HG--
extra : moz-landing-system : lando