testing: remove all unnecessary `hghave` files (Bug 1562960) r=smacleod
Rather than having many `hghave` files, each of which is aware of
it's location in the tree, and can find the well-known `hghave.py`
file in `testing/`, this commit teaches the slightly modified
`run-tests.py` from upstream Mercurial about the location of our
custom `hghave` script. As part of this change, `testing/hghave.py`
needs to be made executable.

It's not pretty, but it's better than the current state, and it
fixes some bugs I found when updating the upstream `run-tests.py`.

Differential Revision: https://phabricator.services.mozilla.com/D36704

--HG--
extra : moz-landing-system : lando