py3: use `pycompat.getargspec` everywhere (Bug 1541166) r=smacleod
`inspect.getargspec` has changed under Python 3, and there is a helper
variable in `mercurial.pycompat` that papers over the differences.
This commit switches all instances in version-control-tools to use
`pycompat.getargspec`. We don't bother changing the single occurance
in `pylib/pycodestyle` since that is third party and will likely be
removed in the near future.

Since the `pycompat.getargspec` helper was introduced in Mercurial
4.6, we bump the `minimumhgversion` string to require at least
that version in both `firefoxtree` and `push-to-try`.

Differential Revision: https://phabricator.services.mozilla.com/D36707

--HG--
extra : moz-landing-system : lando