hghooks: move hard-coded string constants to top level (Bug 1515119) r=glob
These strings are not required to be class-level constant variables,
and their presence within the check class has no functional value.
Leaving them in the class requires prefixing `self.` to each use
of the variable. This commit moves all the variables to the top
level and removes many uses of `self`.

Differential Revision: https://phabricator.services.mozilla.com/D38350

--HG--
extra : moz-landing-system : lando