hghooks: rename privileged push check and change magic words (Bug 1515119) r=glob
After consulting with some engineers from other groups, it was
pointed out that the real intent of the magic words and message
is to have the pushing user provide their reason for not using
Lando. The "privileges" required for this are not as important
as having the user acknowledge they are going outside the
required process, and providing their reason for doing so.

This commit changes the name of the check and most terminology
to "Lando Required Check". The true technical details of this
check are to enforce landings via Lando for a set of repositories,
and create a process for circumventing the Lando requirement.
If the user wishes to go around the automated landing process,
they must specify they are pushing "manually", and thus the
magic words have been changed to indicate they should be used
for that purpose.

Differential Revision: https://phabricator.services.mozilla.com/D38349

--HG--
rename : hghooks/mozhghooks/check/privileged_push_check.py => hghooks/mozhghooks/check/lando_required.py
rename : hghooks/tests/test-privileged-push.t => hghooks/tests/test-lando-required.t
extra : moz-landing-system : lando