hghooks: do not fail when `lando_required_repo_list` config is absent (Bug 1515119) r=glob
When changing the name of the "privileged push" hook, I forgot
to update the `docker-hg-ssh` role to modify the correct Ansible
variable in testing. As a result, the variable was missing from
the config altogether and many tests were failing. This is because
we were not checking the return value of `ui.config` before attempting
to call `.split()` on it. This commit adds logic to assert the
config value is usable before creating the list of repos.

Differential Revision: https://phabricator.services.mozilla.com/D38355

--HG--
extra : moz-landing-system : lando